requirements.txt
opencv-python
face_recognition
Flask
requests
python-dotenv
numpy
